User Profile Page ALPHA Working
Built the UI for the UserProfilePage, implemented necessary functions in the Controller modelled on Adam and Josh's  controllers for SignUp and LogIn. Added in resizing for the UPP so it is the same size as the last page. Added a button to the Dashboard to give access to the settings page and added the necessary user data passer functions. At present you can change your details on the UPP page and then it'll take you to the login screen where you can verify those changes have occured.